Fix EXWM buffers not being hidden on workspace switch
When I switch workspaces, EXWM buffers are not correctly hidden, but stays visible and in the way of the buffers I actually want to use. I tracked it down to the `output-{new,old}` objects being equal component wise, but not reference wise in this `cond` clause. Using the less-strict component-wise equality test of `equal` fixed it for me.

I see there are many uses of `eq` in the EXWM source in general. Maybe there are more bugs in here due to the use of too-strict equality tests with `eq`. Just a thing to consider.