Surely grab keyboard when answering about killing unresponsive app #255
When an app is unresponsive, killing the buffer causes yes-or-no question in
minibuffer. But if the buffer is in char mode, we cannot answer the question.
Changing mode to line mode is also blocked. This is probably because y-or-n-p
catches the mode-transition key. So, first change mode to line mode and question
yes-or-no.